requirements.txt
# Web/app
fastapi==0.116.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.7.3

# NeMo / ASR
nemo_toolkit[asr]==2.4.0
cuda-python==12.9.2   # optional: remove if you prefer not to use it

# Audio + diarizer deps (fallback diarizer)
numpy
scipy
librosa
soundfile==0.12.1
scikit-learn

# For denoise_utils.py
pydub
pyloudnorm
noisereduce
aiofiles
